Fix source ref lookup for data tests
Fixes #5662
