+++ kotest-framework/kotest-framework-engine/api/kotest-framework-engine.api
@@ -1605,6 +1605,10 @@ public final class io/kotest/core/source/SourceRef$None : io/kotest/core/source/
 	public static final field INSTANCE Lio/kotest/core/source/SourceRef$None;
 }
 
+public final class io/kotest/core/source/SourceRefUtils {
+	public static final field INSTANCE Lio/kotest/core/source/SourceRefUtils;
+}
+
 public abstract class io/kotest/core/spec/CompositeSpec : io/kotest/core/spec/Spec {
 	public fun <init> ([Lio/kotest/core/factory/TestFactory;)V
 	public fun rootTests ()Ljava/util/List;
