--- kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/test/TestCaseSourceRefTest.kt
@@ -1,32 +1,0 @@
-package com.sksamuel.kotest.engine.test
-
-import io.kotest.common.reflection.bestName
-import io.kotest.core.annotation.EnabledIf
-import io.kotest.core.annotation.LinuxOnlyGithubCondition
-import io.kotest.core.source.SourceRef
-import io.kotest.core.spec.SpecRef.Reference
-import io.kotest.core.spec.style.FunSpec
-import io.kotest.engine.spec.Materializer
-import io.kotest.matchers.collections.shouldContainExactly
-import kotlin.time.Duration.Companion.seconds
-
-@EnabledIf(LinuxOnlyGithubCondition::class)
-class TestCaseSourceRefTest : FunSpec() {
-   init {
-      test("ClassLineSource should include file name and line number") {
-        val spec = MySpecForTestCaseSourceRefTest()
-        val tests = Materializer().materialize(spec, Reference(spec::class, spec::class.bestName()))
-         tests.map { it.source }.shouldContainExactly(
-            SourceRef.ClassLineSource("com.sksamuel.kotest.engine.test.MySpecForTestCaseSourceRefTest", 14),
-            SourceRef.ClassLineSource("com.sksamuel.kotest.engine.test.MySpecForTestCaseSourceRefTest", 17),
-         )
-      }
-
-      test("source ref should be performant").config(timeout = 20.seconds) {
-         repeat(5_000) {
-           val spec = MySpecForTestCaseSourceRefTest()
-           Materializer().materialize(spec, Reference(spec::class, spec::class.bestName())).first().source
-         }
-      }
-   }
-}
